requirements.txt
cachetools==5.3.3
celery==5.3.6
cloudinary==1.38.0
Deprecated==1.2.14
gunicorn==21.2.0
python-slugify==8.0.3
python-telegram-bot==21.3
rich==13.7.0
six==1.16.0